Convert all timestamps from float to double
The timestamps are long, they really need to be a double!
Without this, we had multiple timestamps that mapped to the same value in
seconds, which meant that our display failed.

The changes are primarily in android, since the iOS code has always been in doubles.
However, while making the change, I noticed the write and read ts columns were
labelled as INTEGER.

This still worked on iOS and works on Android due to some magic with type affinities,
but I change it just in case the magic stops working later.

That is the only iOS change - changing the schema for the table in the blank userCacheDb

+++ android/app/src/main/java/edu/berkeley/eecs/cfc_tracker/location/LocationChangeIntentService.java
@@ -104,7 +104,7 @@ public class LocationChangeIntentService extends IntentService {
         SimpleLocation[] last10Points = uc.getLastSensorData(R.string.key_usercache_filtered_location,
 				FIVE_MINUTES_IN_MS / LocationTrackingConfig.getConfig(this).getDetectionInterval(), SimpleLocation.class);
 
-        float nowSecs = ((float)System.currentTimeMillis())/1000;
+        double nowSecs = ((double)System.currentTimeMillis())/1000;
         UserCache.TimeQuery tq = new UserCache.TimeQuery(R.string.metadata_usercache_write_ts,
                 nowSecs - FIVE_MINUTES_IN_MS - 10, nowSecs);
